updating blurb about hostkey verification in docs, consolidating relavent tests from test_remote_server_key file to test_connection, reworking lookup logic to be in parity with underlying paramiko library, commenting out non-default port tests as pytest-sftpserver currently doesn't support setting an alternate port
