+++ sftpserver.pub
@@ -1,4 +1,4 @@
-localhost ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAID7ImMoiNi/Ncxf2m3nwaupR6wE/vIe5lUSTFkKLzta7
-127.0.0.1 ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAID7ImMoiNi/Ncxf2m3nwaupR6wE/vIe5lUSTFkKLzta7
-|1|JBn58acuz5/x1HJG2fPApHFLidI=|Gb01I0ZsrnY1Hx8/QSrk7ml30pE= ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAID7ImMoiNi/Ncxf2m3nwaupR6wE/vIe5lUSTFkKLzta7
-|1|g5oQ97o5cK0T93sM31Mnxf3qzCo=|kvtLgrmEUWtHvwFPSR04soFNcGg= ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAID7ImMoiNi/Ncxf2m3nwaupR6wE/vIe5lUSTFkKLzta7
+localhost ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCx/Xotg28/q8BzFmFK54MLEimBRs6vgePp/fsH+cBT4QhtT7eLng1dMya0OMb7HNE04fXXSHejs/cDkHtwW74iUbQAfOWwFt878kAbNwtu0qfEPCf8LmAWlQ0paq+P2SCpGauao8uXtSN9jFlVRJEPwaH/Mg429OXcLT8Sw+f/vt7Rgf/ALtAPKDiuKQr0KSx6Sn0+KySCR6oIglOTQROIYaE44NQE5auEcJkFcC9tuEF78RV2JTQhQpTzr0KLR0oo/eHpiN4OE3CuQ1ruNmdc9g9Ed2yZaL5ErvR2yhl78yQrYYTK5yOyHyAIda+W2Uns1AnR+StysR1Onkr111UB
+127.0.0.1 ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCx/Xotg28/q8BzFmFK54MLEimBRs6vgePp/fsH+cBT4QhtT7eLng1dMya0OMb7HNE04fXXSHejs/cDkHtwW74iUbQAfOWwFt878kAbNwtu0qfEPCf8LmAWlQ0paq+P2SCpGauao8uXtSN9jFlVRJEPwaH/Mg429OXcLT8Sw+f/vt7Rgf/ALtAPKDiuKQr0KSx6Sn0+KySCR6oIglOTQROIYaE44NQE5auEcJkFcC9tuEF78RV2JTQhQpTzr0KLR0oo/eHpiN4OE3CuQ1ruNmdc9g9Ed2yZaL5ErvR2yhl78yQrYYTK5yOyHyAIda+W2Uns1AnR+StysR1Onkr111UB
+|1|JBn58acuz5/x1HJG2fPApHFLidI=|Gb01I0ZsrnY1Hx8/QSrk7ml30pE= ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCx/Xotg28/q8BzFmFK54MLEimBRs6vgePp/fsH+cBT4QhtT7eLng1dMya0OMb7HNE04fXXSHejs/cDkHtwW74iUbQAfOWwFt878kAbNwtu0qfEPCf8LmAWlQ0paq+P2SCpGauao8uXtSN9jFlVRJEPwaH/Mg429OXcLT8Sw+f/vt7Rgf/ALtAPKDiuKQr0KSx6Sn0+KySCR6oIglOTQROIYaE44NQE5auEcJkFcC9tuEF78RV2JTQhQpTzr0KLR0oo/eHpiN4OE3CuQ1ruNmdc9g9Ed2yZaL5ErvR2yhl78yQrYYTK5yOyHyAIda+W2Uns1AnR+StysR1Onkr111UB
+|1|g5oQ97o5cK0T93sM31Mnxf3qzCo=|kvtLgrmEUWtHvwFPSR04soFNcGg= ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCx/Xotg28/q8BzFmFK54MLEimBRs6vgePp/fsH+cBT4QhtT7eLng1dMya0OMb7HNE04fXXSHejs/cDkHtwW74iUbQAfOWwFt878kAbNwtu0qfEPCf8LmAWlQ0paq+P2SCpGauao8uXtSN9jFlVRJEPwaH/Mg429OXcLT8Sw+f/vt7Rgf/ALtAPKDiuKQr0KSx6Sn0+KySCR6oIglOTQROIYaE44NQE5auEcJkFcC9tuEF78RV2JTQhQpTzr0KLR0oo/eHpiN4OE3CuQ1ruNmdc9g9Ed2yZaL5ErvR2yhl78yQrYYTK5yOyHyAIda+W2Uns1AnR+StysR1Onkr111UB
